+++ database/migrations/2022_03_02_095704_change_double_price_table.php
@@ -14,7 +14,7 @@ class ChangeDoublePriceTable extends Migration
     public function up()
     {
         Schema::table('plugins_products', function (Blueprint $table) {
-            $table->float('price', 8, 3)->change();
+            $table->double('price')->change();
         });
     }
